color_palette_extraction/requirements.txt
Pillow==9.3.0
selenium==4.6.1
colorthief==0.2.1